pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ev-tripchain"
version = "0.1.0"
description = "EV trip-chain charging load generation + probabilistic hosting capacity assessment for distribution grids"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pydantic>=2.7",
  "pyyaml>=6.0.1",
  "typer>=0.12",
  "pandapower>=2.14",
  "matplotlib>=3.8",
  "numba>=0.64.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6"]

[project.scripts]
ev-tripchain = "ev_tripchain.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ev_tripchain"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
